[tz-307] Repopulate validation editor when editing existing form element with validation validators.
