+++ modules/forms/src/lib/components/form-element-editor/form-element-editor.component.html
@@ -27,7 +27,7 @@
 
     <classifieds-ui-datasource-options formControlName="datasourceOptions" [datasourceOptions]="datasourceOptions"></classifieds-ui-datasource-options>
 
-    <druid-ordain-validation-editor formControlName="validation"></druid-ordain-validation-editor>
+    <druid-ordain-validation-editor formControlName="validation" [validation]="validation"></druid-ordain-validation-editor>
 
   </mat-dialog-content>
   <mat-dialog-actions>
